Fix: 디자인 사항들 수정 및 프로필 페이지 네비게이션 추가
